Add supervisor package for managing service lifetimes
A supervisor is essentially a more capable errgroup. It monitors a set
of running services, and restarts them if they fail.

The supervisor keeps track of the status of each service and reports any
status changes to listeners via a callback.

The supervisor API is designed to be compatible with our wire-based
initialization flow, where any service that wants to run in the
supervisor should depend on (and add themselves to) a supervisor.Config.

This initial version of the supervisor is intended to be as simple as
possible, making a number of assumptions:

* All services should run for the entire lifetime of the supervisor
* A service that exits gracefully before the supervisor is shut down
will be restarted